Fixed issues when plugin file was guessed by slug.
If the plugin file was guessed by slug the `findMainFile` would return
only the (relative) plugin filename, if it was found by globbing the
PHP files in the main directory it returns a full absolute path to the
main plugin file. This resulted in a bugs when the main plugin file was
guessed by slug.